ODK-32: Added export of data from ODK tables to CSV and JSON files
Added export functionality which enables downloading data from a particular table to both CSV and JSON files. There are now dedicated buttons for exporting in the Current Tables tab which trigger a popup which let users decide if they want an export of the data to the CSV or JSON file

Note: Data in the CSV files is separated by tab key, so tab should be used as a separator when opening the CSV file in a dedicated tool like Excel.